Unify metrics for blob size/history size/history count
